Delete type member function from tetromino interface
It doesn't seem to be a good design to have an interface method to
identify the type of the class implementing it. Users of the interface
shouldn't have to need this information: either the interface is
missing something or the design is not correct.

In this case, a playfield object only has to know the color of the
tetromino occupying a certain cell, not its type.

This commit just fixes the design of the tetromino interface and
modifies whatever takes so the program compiles and the game is still
playable. Note that stacked tetrominoes are not coloured. This will be
fixed in following commits.