Restore color to stacked tetrominoes within the playfield
The playfield is defined now as a vector of optional tetrions, meaning
that each cell is occupied by one of them or is empty. Note that no
tetrion type has been defined, as we're just interested in its color.
That's why the well has been declared as a list of optional colors.